Se modifico el metos eliminar qunque aun falla

+++ backend/src/main/java/mx/uam/ayd/proyecto/negocio/ServicioAsesoria.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import lombok.extern.slf4j.Slf4j;
 import mx.uam.ayd.proyecto.datos.AlumnoRepository;
 import mx.uam.ayd.proyecto.datos.AsesoriaRepository;
 import mx.uam.ayd.proyecto.datos.MateriaRepository;
@@ -12,7 +13,9 @@ import mx.uam.ayd.proyecto.dto.AsesoriaDto;
 import mx.uam.ayd.proyecto.negocio.modelo.Alumno;
 import mx.uam.ayd.proyecto.negocio.modelo.Asesoria;
 import mx.uam.ayd.proyecto.negocio.modelo.Materia;
+import mx.uam.ayd.proyecto.servicios.AsesoriaRestController;
 @Service
+@Slf4j
 public class ServicioAsesoria {
 	
 	@Autowired 
@@ -23,6 +26,7 @@ public class ServicioAsesoria {
 	
 	@Autowired 
 	private MateriaRepository materiaRepository;
+	
 	/**
 	 * Agregar una nueva asesoria
 	 * 
@@ -40,7 +44,6 @@ public class ServicioAsesoria {
 		}
 		
 		Alumno alumno = optAlumno.get();
-		
 		Optional<Materia> optMateria = materiaRepository.findById(asesoriaDto.getMateria());
 		
 		if(optMateria.isEmpty()) {
@@ -58,20 +61,23 @@ public class ServicioAsesoria {
 		asesoria.setHoraTermino(asesoriaDto.getHoraTermino());
 		asesoria.setCosto(asesoriaDto.getCosto());
 		asesoria.setUbicacion(asesoriaDto.getUbicacion());
-		asesoria.setUrl(asesoriaDto.getUrl());
-		asesoria.setTotalPuntuaciones(asesoriaDto.getTotalPuntuaciones());
-		asesoria.setPuntuacion(asesoriaDto.getPuntuacion());
-		asesoria.setEstado(asesoriaDto.getEstado());
 		asesoria.setMateria(materia);
 		asesoria.setIdAlumno(alumno.getIdAlumno());
-		asesoria = asesoriaRepository.save(asesoria);
 		
+		for(Asesoria asesorias: asesoriaRepository.findAll()) {
+			if((id == asesorias.getIdAlumno()) & (asesoriaDto.getDia().equals(asesorias.getDia())) & (asesoriaDto.getHoraInicio().equals(asesorias.getHoraInicio())) & (asesoriaDto.getHoraTermino().equals(asesorias.getHoraTermino()))) {
+				throw new IllegalArgumentException("No se puede repetir");
+			} 
+		}
+
+		asesoria = asesoriaRepository.save(asesoria);
+	
 		materia.addAsesoria(asesoria);
 		materiaRepository.save(materia);
 		
 		alumno.addAsesoria(asesoria);
 		alumnoRepository.save(alumno);
-		
+	
 		return AsesoriaDto.creaAsesoriaDto(asesoria);
 	}
 	
@@ -79,10 +85,24 @@ public class ServicioAsesoria {
 	 * Se elimina una asesoria con su id
 	 * 
 	 * @param idAsesoria el id de la asesoria
+	 * @param idAlumno el id del usuario
 	 */
-	public void eliminarAsesoria(Long idAsesoria) {
+	public void eliminarAsesoria(Long idAsesoria, Long idAlumno) {
 		
-		asesoriaRepository.deleteById(idAsesoria);
+		Optional<Asesoria> optAsesoria = asesoriaRepository.findById(idAsesoria);
+		if(optAsesoria.isEmpty())
+			throw new IllegalArgumentException("La asessoria no existe");
+		
+		
+		Optional<Alumno> optAlumno = alumnoRepository.findById(idAlumno);
+		if(optAlumno.isEmpty())
+			throw new IllegalArgumentException("El alumno no esta en la BD");
+		
+		if(optAsesoria.get().getIdAlumno() == optAlumno.get().getIdAlumno()) {
+			log.info("Se va a eliminar la asesoria con id: "+ idAsesoria);
+			asesoriaRepository.deleteById(idAsesoria);
+		}
+		
 	}
 	
 	/**
@@ -93,9 +113,7 @@ public class ServicioAsesoria {
 	 */
 	public AsesoriaDto retrieve(Long idAsesoria) {
 		Optional<Asesoria> optAsesoria = asesoriaRepository.findById(idAsesoria);
-		
-		Asesoria asesoria = optAsesoria.get();
-		return AsesoriaDto.creaAsesoriaDto(asesoria);
+		return AsesoriaDto.creaAsesoriaDto(optAsesoria.get());
 	}
 
 }
